Revert "Adding Hiatus Update Page (#37)"
This reverts commit 0abe101c10cd0c9ecdd1756a203d0bcf71486034.
