--- content/hiatus.mdx
@@ -1,8 +1,0 @@
----
-title: "My Hiatus From Work"
-subheader: "*Daily* summaries on what I do each day with no job"
-date: "2024-11-15"
-category: industry
----
-
-<GoogleDoc documentId="1Z3lUrw3VOf7TfE0CEiBEfCNrNKAOu-sQmt0xITShgEw" />
