+++ contentlayer.config.js
@@ -1,4 +1,4 @@
-import { defineDocumentType, makeSource } from "contentlayer2/source-files"
+import { defineDocumentType, makeSource } from "contentlayer/source-files"
 import remarkGfm from "remark-gfm"
 import rehypeKatex from "rehype-katex";
 import remarkMath from "remark-math";
