+++ next.config.js
@@ -1,4 +1,4 @@
-const { withContentlayer } = require("next-contentlayer2")
+const { withContentlayer } = require("next-contentlayer")
 
 /** @type {import('next').NextConfig} */
 let nextConfig = {
