+++ package.json
@@ -7,35 +7,35 @@
         "prod": "next build && next start"
     },
     "dependencies": {
-        "@emotion/react": "^11.13.3",
-        "@emotion/styled": "^11.13.0",
-        "@fontsource/lato": "^5.1.0",
-        "@fontsource/pt-mono": "^5.1.0",
-        "@mui/icons-material": "^6.1.7",
-        "@mui/material": "^6.1.7",
-        "@vercel/analytics": "^1.4.0",
-        "contentlayer2": "^0.5.3",
+        "@emotion/react": "^11.11.0",
+        "@emotion/styled": "^11.11.0",
+        "@fontsource/lato": "^4.5.10",
+        "@fontsource/pt-mono": "^4.5.11",
+        "@mui/icons-material": "^5.11.16",
+        "@mui/material": "^5.12.3",
+        "@vercel/analytics": "^0.1.8",
+        "contentlayer": "^0.3.2",
         "cross-env": "^7.0.3",
-        "date-fns": "^4.1.0",
-        "katex": "^0.16.11",
-        "next": "^15.0.3",
-        "next-contentlayer2": "^0.5.3",
-        "react": "^18.3.1",
-        "react-dom": "^18.3.1",
-        "rehype-highlight": "^7.0.1",
-        "rehype-katex": "^7.0.1",
-        "rehype-pretty-code": "^0.14.0",
-        "rehype-stringify": "^10.0.1",
-        "remark-gfm": "^4.0.0",
-        "remark-math": "^6.0.0",
-        "remark-parse": "^11.0.0",
-        "remark-rehype": "^11.1.1",
-        "shiki": "^1.23.0",
-        "styled-jsx": "^5.1.6",
-        "swr": "^2.2.5",
-        "unified": "^11.0.5"
+        "date-fns": "^2.29.3",
+        "katex": "^0.16.7",
+        "next": "^13.1.6",
+        "next-contentlayer": "^0.3.2",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "rehype-highlight": "^6.0.0",
+        "rehype-katex": "^6.0.3",
+        "rehype-pretty-code": "^0.9.5",
+        "rehype-stringify": "^9.0.3",
+        "remark-gfm": "^3.0.1",
+        "remark-math": "^5.1.1",
+        "remark-parse": "^10.0.1",
+        "remark-rehype": "^10.1.0",
+        "shiki": "^0.14.2",
+        "styled-jsx": "^5.1.2",
+        "swr": "^2.1.5",
+        "unified": "^10.1.2"
     },
     "devDependencies": {
-        "@svgr/webpack": "^8.1.0"
+        "@svgr/webpack": "^7.0.0"
     }
 }
