+++ pages/blog/[category]/[id].js
@@ -9,9 +9,8 @@ import styles from '@/styles/blog/Post.module.css';
 import blogCode from '@/styles/blogCode';
 
 import { allPosts } from '@/.contentlayer/generated';
-import { useMDXComponent } from 'next-contentlayer2/hooks'
+import { useMDXComponent } from 'next-contentlayer/hooks'
 import { ArticleContent, MultiLevelArticleContent } from '@/components/blog/articleContent';
-import GoogleDoc from '@/dynamic-content/googleDoc';
 
 export function getStaticPaths() {
     const paths = allPosts.map((post) => {
@@ -49,7 +48,6 @@ export default function Post({ postData }) {
         a: ({ href, children }) => <Link href={href.toString()}>{children}</Link>,
         img: ({ src, alt }) => <Image src={src} alt={alt} width={1000} height={300} />,
         ArticleContent: postData.multilevelarticle ? MultiLevelArticleContent : ArticleContent,
-        GoogleDoc
     };
 
 
@@ -81,3 +79,4 @@ export default function Post({ postData }) {
         </>
     )
 }
+
